# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Creative Studio CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign with services and milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign with services and milestones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/share:
    post:
      operationId: shareCampaign
      tags: [campaigns]
      summary: Share campaign with client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/unshare:
    post:
      operationId: unshareCampaign
      tags: [campaigns]
      summary: Unshare campaign from client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unshared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /campaigns/{id}/services:
    get:
      operationId: listCampaignServices
      tags: [campaigns]
      summary: List services for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of campaign services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignService"
    post:
      operationId: addCampaignService
      tags: [campaigns]
      summary: Add a service to campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCampaignServiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignService"
  /campaigns/{id}/services/{serviceId}:
    delete:
      operationId: removeCampaignService
      tags: [campaigns]
      summary: Remove a service from campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/milestones:
    get:
      operationId: listCampaignMilestones
      tags: [campaigns]
      summary: List milestones for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignMilestone"
    post:
      operationId: addCampaignMilestone
      tags: [campaigns]
      summary: Add a milestone to campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCampaignMilestoneBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignMilestone"
  /campaigns/{id}/milestones/{milestoneId}:
    patch:
      operationId: updateCampaignMilestone
      tags: [campaigns]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignMilestoneBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignMilestone"
    delete:
      operationId: deleteCampaignMilestone
      tags: [campaigns]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        nameFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        descriptionFr:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        proposalContent:
          type: ["string", "null"]
        proposalContentAr:
          type: ["string", "null"]
        proposalContentFr:
          type: ["string", "null"]
        shared:
          type: boolean
        sharedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - status
        - shared
        - createdAt
        - updatedAt
    CampaignDetail:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/CampaignService"
            milestones:
              type: array
              items:
                $ref: "#/components/schemas/CampaignMilestone"
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        nameFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        descriptionFr:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        clientId:
          type: ["integer", "null"]
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        proposalContent:
          type: ["string", "null"]
        proposalContentAr:
          type: ["string", "null"]
        proposalContentFr:
          type: ["string", "null"]
      required:
        - name
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        nameAr:
          type: ["string", "null"]
        nameFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        descriptionFr:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        proposalContent:
          type: ["string", "null"]
        proposalContentAr:
          type: ["string", "null"]
        proposalContentFr:
          type: ["string", "null"]
        shared:
          type: ["boolean", "null"]
    CampaignService:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        serviceId:
          type: integer
        customPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    AddCampaignServiceBody:
      type: object
      properties:
        serviceId:
          type: integer
        customPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - serviceId
    CampaignMilestone:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        title:
          type: string
        titleAr:
          type: ["string", "null"]
        titleFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
        dueDate:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - campaignId
        - title
        - order
        - isCompleted
        - createdAt
    AddCampaignMilestoneBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: ["string", "null"]
        titleFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
    UpdateCampaignMilestoneBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        titleAr:
          type: ["string", "null"]
        titleFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        isCompleted:
          type: ["boolean", "null"]
...